Changes to Listing Review DTO Add queries for ListingReview Listing Review Endpoint and Service to add reviews and view host reviews Add reserve method to Listing Service & Endpoint

+++ src/main/java/com/bookaroom/entities/ListingReviewDTO.java
@@ -25,7 +25,7 @@ public class ListingReviewDTO
     private Long userId;
 
     @Column(name = "RATING")
-    private Integer rating;
+    private Float rating;
 
     @Column(name = "COMMENTS")
     private String comments;
@@ -63,12 +63,12 @@ public class ListingReviewDTO
         this.userId = userId;
     }
 
-    public Integer getRating()
+    public Float getRating()
     {
         return rating;
     }
 
-    public void setRating(Integer rating)
+    public void setRating(Float rating)
     {
         this.rating = rating;
     }
